fix test; compile with static lib

+++ paddle/phi/kernels/funcs/warp_transducer/tests/random.cpp
@@ -26,7 +26,7 @@ float* genActs(int size) {
 void genActs(std::vector<float>& arr) {
   std::mt19937 gen(0);
   std::uniform_real_distribution<> dis(0, 1);
-  for (int i = 0; i < arr.size(); ++i) arr[i] = dis(gen);
+  for (size_t i = 0; i < arr.size(); ++i) arr[i] = dis(gen);
 }
 
 std::vector<int> genLabels(int alphabet_size, int L) {
